# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Productos - Laravel 11 con Swagger",
        "description": "Esta API permite registrar usuarios, iniciar sesión, gestionar productos (CRUD) y consultar productos externos.",
        "contact": {
            "name": "Andrés",
            "email": "andres@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar un nuevo usuario",
                "description": "Crea un nuevo usuario y devuelve un token de acceso.",
                "operationId": "f7fb8560722ba1f43f9e570fe9affddb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Andrés"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "andres@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|Xyz123TokenEjemplo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "description": "Valida las credenciales y devuelve un token de acceso.",
                "operationId": "1257f11ade1e6ee3a26cfe338badfb8e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "andres@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|Xyz123TokenEjemplo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "description": "Invalida el token activo del usuario autenticado.",
                "operationId": "d0090497298d35d96a9fcde3666ff010",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sesión cerrada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/external-products": {
            "get": {
                "tags": [
                    "Productos Externos"
                ],
                "summary": "Obtener productos desde una API externa",
                "operationId": "b0c2e4a7db9eef2ca2102a7126cf9fda",
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida exitosamente"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener todos los productos",
                "operationId": "7cd9d394f726672223f216b77acab086",
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida exitosamente"
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "e83af3dd12f9ebc4ad755304893c016f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Camiseta Nike"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Camiseta deportiva de algodón"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "59.99"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener un producto por ID",
                "operationId": "fa07b5c9bbd4d51f1f3763ca56e0908b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a buscar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto encontrado exitosamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto existente",
                "operationId": "e0e4bcc68686fa05279e23580e41bea1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Zapatos Adidas"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Zapatos deportivos para correr"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "79.99"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado exitosamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto por ID",
                "operationId": "bae80a1b7d88ec2b8f025a048aa00f86",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado correctamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            }
        },
        "/api/external-products": {
            "get": {
                "tags": [
                    "Productos Externos"
                ],
                "summary": "Obtener productos desde una API externa",
                "description": "Consume datos desde la API pública fakestoreapi.com y los devuelve en formato JSON.",
                "operationId": "d5102a1c6073fe0b4fc09d5c2a920009",
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida exitosamente desde la API externa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Fjallraven - Foldsack No. 1 Backpack"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 109.95
                                            },
                                            "category": {
                                                "type": "string",
                                                "example": "men's clothing"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Mochila perfecta para uso diario y actividades al aire libre"
                                            },
                                            "image": {
                                                "type": "string",
                                                "example": "https://fakestoreapi.com/img/81fPKd-2AYL._AC_SL1500_.jpg"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al conectarse con la API externa"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Autenticación",
            "description": "Registro, inicio de sesión y cierre de sesión."
        },
        {
            "name": "Productos Externos",
            "description": "Productos Externos"
        },
        {
            "name": "Productos",
            "description": "Productos"
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Ingresa token: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}